Rewrite and rename `Match.OneByRegex()`
This function is now called `Match.OneByRegexOrPrefix()` and has been
rewrite to use a `List<(string, string, string)>` instead of the
original `Dictionary<string, string>`.
The goal of that is to prepare this fonction (called only to find a well
known alias for the `Sid` class) to handle a list of tuple wich contains
both bigram (when applicable) and the regex SID associated with the
alias name.