Fix ACE's equality
This need to be fixed since as the doc says
(https://docs.microsoft.com/en-us/dotnet/api/system.linq.enumerable.except?view=netframework-4.8)
the `except()` method output a collection of things that are present in
the first and not in not second collection but if we have something in
the second collection which is not in the first one, the `except()` method
returns nothing at all.
Following that, the implementation of the ACE's equality is wrong

This commit fix that by adding a second call to the `except()` method
but from the compared object instead of the comparer one.